backend/requirements.txt
web3==6.20.0
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
numpy==2.0.1
matplotlib==3.8.4
